# openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "blueeconomy-insurance",
  "version": "0.1.0"
 },
 "paths": {
...
  "/v1/products": {
   "get": {
    "tags": [
     "products"
    ],
    "summary": "List Products",
    "operationId": "list_products_v1_products_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response List Products V1 Products Get"
        }
       }
      }
     }
    }
   },
   "post": {
    "tags": [
     "products"
    ],
    "summary": "Create Product",
    "operationId": "create_product_v1_products_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/ProductCreate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "type": "object",
         "title": "Response Create Product V1 Products Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/products/{code}/versions/{version}/activate": {
   "post": {
    "tags": [
     "products"
    ],
    "summary": "Activate Product",
    "operationId": "activate_product_v1_products__code__versions__version__activate_post",
    "parameters": [
     {
      "name": "code",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Code"
      }
     },
     {
      "name": "version",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Version"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Activate Product V1 Products  Code  Versions  Version  Activate Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/v1/products/{code}/versions/{version}/rate-tables": {
   "post": {
    "tags": [
     "products"
    ],
    "summary": "Add Rate Table",
    "operationId": "add_rate_table_v1_products__code__versions__version__rate_tables_post",
    "parameters": [
     {
      "name": "code",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Code"
      }
     },
     {
      "name": "version",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Version"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RateTableCreate"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response Add Rate Table V1 Products  Code  Versions  Version  Rate Tables Post"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "products"
    ],
    "summary": "List Rate Tables",
    "operationId": "list_rate_tables_v1_products__code__versions__version__rate_tables_get",
    "parameters": [
     {
      "name": "code",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Code"
      }
     },
     {
      "name": "version",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "Version"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "object",
         "additionalProperties": true,
         "title": "Response List Rate Tables V1 Products  Code  Versions  Version  Rate Tables Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "ProductCreate": {
    "properties": {
     "code": {
      "type": "string",
      "pattern": "^[a-z0-9][a-z0-9._-]{0,63}$",
      "title": "Code"
     },
     "kind": {
      "type": "string",
      "enum": [
       "marine-cargo-single",
       "marine-cargo-open",
       "ferry-parametric",
       "protection-indemnity",
       "hull"
      ],
      "title": "Kind"
     },
     "name": {
      "type": "string",
      "maxLength": 256,
      "minLength": 1,
      "title": "Name"
     },
     "definition": {
      "additionalProperties": true,
      "type": "object",
      "title": "Definition"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "code",
     "kind",
     "name"
    ],
    "title": "ProductCreate"
   },
...
   "RateTableCreate": {
    "properties": {
     "effective_from": {
      "type": "string",
      "format": "date",
      "title": "Effective From"
     },
     "effective_to": {
      "anyOf": [
       {
        "type": "string",
        "format": "date"
       },
       {
        "type": "null"
       }
      ],
      "title": "Effective To"
     },
     "rates": {
      "additionalProperties": true,
      "type": "object",
      "title": "Rates"
     }
    },
    "additionalProperties": false,
    "type": "object",
    "required": [
     "effective_from",
     "rates"
    ],
    "title": "RateTableCreate"
   },
...
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   }
  }
 }
}